deleteLabData: Added logic to delete lab's blob @close #66

+++ src/app/api/provider/deleteLabData/route.js
@@ -1,6 +1,7 @@
 import path from 'path';
 import { promises as fs } from 'fs';
 import { NextResponse } from 'next/server';
+import { del } from '@vercel/blob';
 
 export async function POST(req) {
   try {
@@ -8,20 +9,39 @@ export async function POST(req) {
     if (!labURI) {
       return NextResponse.json({ error: 'Missing labURI' }, { status: 400 });
     }
-
+    const isVercel = !!process.env.VERCEL;
     const filePath = path.join(process.cwd(), 'data', labURI);
 
-    try {
-      await fs.unlink(filePath);
-      console.log(`Deleted file: ${filePath}`);
-      return NextResponse.json({ message: 'Lab data deleted successfully.' }, { status: 200 });
-    } catch (error) {
-      console.error("Error deleting file:", error);
-      return NextResponse.json(
-        { error: 'Failed to delete lab data.', details: error.message },
-        { status: 500 }
-      );
+    if (!isVercel) {
+        try {
+        await fs.unlink(filePath);
+        return NextResponse.json({ message: 'Lab data deleted successfully.' }, { status: 200 });
+        } catch (error) {
+        return NextResponse.json(
+            { error: 'Failed to delete lab data.', details: error.message },
+            { status: 500 }
+        );
+        }
+    } else {
+        try {
+            const blobUrl = `${process.env.VERCEL_BLOB_BASE_URL}/data/${labURI}`;
+            const result = await del(blobUrl);
+            if (result) {
+            console.log(`Blob deleted from Vercel: ${blobUrl}`);
+            return NextResponse.json({ message: 'Lab data deleted successfully.' }, { status: 200 }); //success
+            } else {
+            console.warn(`Blob deletion from Vercel may have failed: ${blobUrl}`); // No falla si no existe.
+            return NextResponse.json({ message: 'Lab data deleted successfully.' }, { status: 200 }); //success
+            }
+        } catch (error) {
+            console.error("Error deleting data:", error);
+            return NextResponse.json(
+                { error: 'Failed to delete lab data.', details: error.message },
+                { status: 500 }
+            );
+        }
     }
+
   } catch (error) {
     return NextResponse.json({ error: 'Failed to parse request body.', details: error.message }, { status: 400 });
   }
